# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ana Portfolio API
  description: >-
    API do portfólio: conteúdo público, editor autenticado, aparência e
    assistente.


    ### Como usar (passo a passo)

    1. Teste a API em **GET /api/v1/health**.

    2. Leia dados com os **GET** 🔓 (públicos) e anote o campo `revision`.

    3. Para alterar dados, faça login em **POST /api/v1/auth/login** e copie o
    `accessToken`.

    4. Clique em **Authorize** (topo da página), cole o token e confirme.

    5. Use os endpoints 🔒 (POST, PUT, DELETE) enviando `expectedRevision`: a
    **versão atual dos dados**, igual ao `revision` do último GET.


    ### Erros comuns

    - **400**: campo faltando ou em formato errado.

    - **401**: sem login ou token vencido.

    - **404**: o `id` não existe.

    - **409**: `expectedRevision` desatualizado. Faça um novo GET e tente de
    novo.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    def-5:
      type: object
      description: Um projeto exibido no portfólio.
      required:
        - id
        - name
        - description
        - stack
        - href
        - status
        - visibility
        - linked
      additionalProperties: false
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 100
          description: 'Identificador único do projeto, em texto simples (ex.:
            "empresa-exemplo"). Não pode repetir.'
          example: exemplo-1
        name:
          type: string
          minLength: 1
          maxLength: 160
          description: Nome do projeto.
          example: Projeto Exemplo
        description:
          type: string
          minLength: 1
          maxLength: 3000
          description: O que o projeto faz.
          example: Aplicação de demonstração com testes automatizados.
        stack:
          type: array
          maxItems: 50
          description: Tecnologias usadas.
          items:
            type: string
            minLength: 1
            maxLength: 80
          example:
            - React
            - TypeScript
            - Fastify
        href:
          type: string
          format: uri
          description: Link do projeto (repositório ou site).
          example: https://github.com/exemplo/projeto-exemplo
        status:
          type: string
          minLength: 1
          maxLength: 80
          description: Situação do projeto.
          example: Em andamento
        visibility:
          type: string
          enum:
            - public
            - private
          description: "`public` aparece para todos; `private` fica restrito."
          example: public
        linked:
          type: boolean
          description: "`true` se o projeto tem link clicável."
          example: true
      title: Project
      example:
        id: projeto-exemplo
        name: Projeto Exemplo
        description: Aplicação de demonstração com testes automatizados.
        stack:
          - React
          - TypeScript
          - Fastify
        href: https://github.com/exemplo/projeto-exemplo
        status: Em andamento
        visibility: public
        linked: true
...
    def-18:
      type: object
      description: Portfólio completo + controle de versão.
      required:
        - revision
        - updatedAt
        - data
      properties:
        revision:
          type: integer
          minimum: 1
          description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
          example: 5
        updatedAt:
          type: string
          format: date-time
          description: Data da última alteração.
          example: 2025-01-15T12:00:00.000Z
        data:
          $ref: "#/components/schemas/def-17"
      title: PortfolioSnapshot
...
paths:
...
  /api/v1/projects:
    get:
      summary: Lista os projetos
      tags:
        - Projects
      description: >-
        🔓 **Público** — não precisa de login.


        Retorna todos os projetos na ordem em que aparecem no portfólio. Guarde
        o valor `revision`: ele é usado nas escritas.
      responses:
        "200":
          description: Lista de projetos.
          content:
            application/json:
              schema:
                description: Lista de projetos.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/def-5"
                    example:
                      - id: projeto-exemplo
                        name: Projeto Exemplo
                        description: Aplicação de demonstração com testes automatizados.
                        stack:
                          - React
                          - TypeScript
                          - Fastify
                        href: https://github.com/exemplo/projeto-exemplo
                        status: Em andamento
                        visibility: public
                        linked: true
    post:
      summary: Cria um projeto
      tags:
        - Projects
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Cria um novo projeto no portfólio. Antes de enviar, consulte o GET
        correspondente e use o valor `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expectedRevision
                - data
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                data:
                  $ref: "#/components/schemas/def-5"
            example:
              expectedRevision: 4
              data:
                id: projeto-exemplo
                name: Projeto Exemplo
                description: Aplicação de demonstração com testes automatizados.
                stack:
                  - React
                  - TypeScript
                  - Fastify
                href: https://github.com/exemplo/projeto-exemplo
                status: Em andamento
                visibility: public
                linked: true
      security: *a1
      responses:
        "201":
          description: Criada. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/projects/{id}:
    get:
      summary: Busca um projeto pelo id
      tags:
        - Projects
      description: |-
        🔓 **Público** — não precisa de login.

        Retorna um único projeto.
      parameters:
        - schema:
            type: string
            minLength: 1
          example: projeto-exemplo
          in: path
          name: id
          required: true
          description: ID do projeto. Use o campo `id` retornado na listagem.
      responses:
        "200":
          description: O projeto encontrado.
          content:
            application/json:
              schema:
                description: O projeto encontrado.
                type: object
                required:
                  - revision
                  - updatedAt
                  - data
                properties:
                  revision:
                    type: integer
                    description: Versão atual dos dados. Use em `expectedRevision` ao salvar.
                    example: 4
                  updatedAt:
                    type: string
                    format: date-time
                    description: Data da última alteração.
                    example: 2025-01-15T12:00:00.000Z
                  data:
                    $ref: "#/components/schemas/def-5"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "404":
          description: "Não encontrado: o `id` informado não existe."
          content:
            application/json:
              schema:
                description: "Não encontrado: o `id` informado não existe."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: PROJECT_NOT_FOUND
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Projeto não encontrado.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      id: empresa-exemplo
    put:
      summary: Atualiza um projeto
      tags:
        - Projects
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Troca os dados de um projeto existente. O `id` da URL deve ser igual ao
        `data.id` do corpo. Antes de enviar, consulte o GET correspondente e use
        o valor `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expectedRevision
                - data
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                data:
                  $ref: "#/components/schemas/def-5"
            example:
              expectedRevision: 4
              data:
                id: projeto-exemplo
                name: Projeto Exemplo
                description: Aplicação de demonstração com testes automatizados.
                stack:
                  - React
                  - TypeScript
                  - Fastify
                href: https://github.com/exemplo/projeto-exemplo
                status: Em andamento
                visibility: public
                linked: true
      parameters:
        - schema:
            type: string
            minLength: 1
          example: projeto-exemplo
          in: path
          name: id
          required: true
          description: ID do projeto. Use o campo `id` retornado na listagem.
      security: *a1
      responses:
        "200":
          description: Salvo. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "404":
          description: "Não encontrado: o `id` informado não existe."
          content:
            application/json:
              schema:
                description: "Não encontrado: o `id` informado não existe."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: PROJECT_NOT_FOUND
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Projeto não encontrado.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      id: empresa-exemplo
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
    delete:
      summary: Remove um projeto
      tags:
        - Projects
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Apaga um projeto. Não tem corpo: envie `expectedRevision` na URL (ex.:
        `?expectedRevision=4`). Antes de enviar, consulte o GET correspondente e
        use o valor `revision` retornado em `expectedRevision`.
      parameters:
        - schema:
            type: integer
            minimum: 1
          example: 4
          in: query
          name: expectedRevision
          required: true
          description: "Versão atual dos dados. Use o valor `revision` retornado pelo
            último GET (vai na URL, ex.: `?expectedRevision=4`)."
        - schema:
            type: string
            minLength: 1
          example: projeto-exemplo
          in: path
          name: id
          required: true
          description: ID do projeto. Use o campo `id` retornado na listagem.
      security: *a1
      responses:
        "200":
          description: Removida. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "404":
          description: "Não encontrado: o `id` informado não existe."
          content:
            application/json:
              schema:
                description: "Não encontrado: o `id` informado não existe."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: PROJECT_NOT_FOUND
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Projeto não encontrado.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      id: empresa-exemplo
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
  /api/v1/projects/order:
    put:
      summary: Reordena os projetos
      tags:
        - Projects
      description: >-
        🔒 **Requer Bearer Token** — faça login em `POST /api/v1/auth/login` e
        cole o `accessToken` no botão **Authorize**.


        Define a nova ordem dos projetos. Envie todos os `id` atuais, sem
        repetir e sem faltar nenhum. Antes de enviar, consulte o GET
        correspondente e use o valor `revision` retornado em `expectedRevision`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Nova ordem. A lista deve ter exatamente os IDs atuais, sem repetir.
              required:
                - expectedRevision
                - ids
              additionalProperties: false
              properties:
                expectedRevision:
                  type: integer
                  minimum: 1
                  description: Versão atual dos dados. Use o valor `revision` retornado pelo
                    último GET.
                  example: 4
                ids:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  description: IDs na ordem desejada (o primeiro aparece primeiro).
                  items:
                    type: string
                    minLength: 1
                  example:
                    - projeto-b
                    - projeto-a
            example:
              expectedRevision: 4
              ids:
                - projeto-b
                - projeto-a
        description: Nova ordem. A lista deve ter exatamente os IDs atuais, sem repetir.
      security: *a1
      responses:
        "200":
          description: Reordenada. Retorna o portfólio completo com a nova `revision`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/def-18"
        "400":
          description: "Dados inválidos: algum campo está faltando ou fora do formato."
          content:
            application/json:
              schema:
                description: "Dados inválidos: algum campo está faltando ou fora do formato."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: VALIDATION_ERROR
                  message:
                    type: string
                    description: Explicação do erro.
                    example: A requisição não atende ao contrato da API.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "401":
          description: "Sem login: envie o Bearer Token (botão Authorize)."
          content:
            application/json:
              schema:
                description: "Sem login: envie o Bearer Token (botão Authorize)."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: UNAUTHORIZED
                  message:
                    type: string
                    description: Explicação do erro.
                    example: Autenticação obrigatória ou token inválido.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
        "409":
          description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
            existe). Faça um novo GET e tente de novo."
          content:
            application/json:
              schema:
                description: "Conflito: `expectedRevision` está desatualizado (ou o `id` já
                  existe). Faça um novo GET e tente de novo."
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    description: Código do erro.
                    example: REVISION_CONFLICT
                  message:
                    type: string
                    description: Explicação do erro.
                    example: O portfólio foi alterado depois que esta edição começou. Recarregue a
                      revisão atual antes de salvar.
                  details:
                    type: object
                    additionalProperties: true
                    description: Informações extras (opcional).
                    example:
                      expectedRevision: 3
                      currentRevision: 4
...
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
...